Add aspect ratio detection for amcodec h264
It is using same approach as amcodec is using for MPEG2 streams.
Skipped frame rate calculation at the moment, although necessary values are available for that.

Also I noticed that there is already a function for SPS parsing,
maybe it is worth to improve that function and have only one function.